Mejoras en BD

+++ src/java/pojo/Examenasisten.java
@@ -1,6 +1,9 @@
 package pojo;
 // Generated Oct 18, 2011 12:37:12 AM by Hibernate Tools 3.2.1.GA
 
+import java.math.BigDecimal;
+
+
 
 
 /**
@@ -12,7 +15,7 @@ public class Examenasisten  implements java.io.Serializable {
      private int id;
      private Preinscrito preinscrito;
      private Examen examen;
-     private Float notaobtenida;
+     private BigDecimal notaobtenida;
      private Boolean borrado;
 
     public Examenasisten() {
@@ -22,7 +25,7 @@ public class Examenasisten  implements java.io.Serializable {
     public Examenasisten(int id) {
         this.id = id;
     }
-    public Examenasisten(int id, Preinscrito preinscrito, Examen examen, Float notaobtenida, Boolean borrado) {
+    public Examenasisten(int id, Preinscrito preinscrito, Examen examen, BigDecimal notaobtenida, Boolean borrado) {
        this.id = id;
        this.preinscrito = preinscrito;
        this.examen = examen;
@@ -51,11 +54,11 @@ public class Examenasisten  implements java.io.Serializable {
     public void setExamen(Examen examen) {
         this.examen = examen;
     }
-    public Float getNotaobtenida() {
+    public BigDecimal getNotaobtenida() {
         return this.notaobtenida;
     }
     
-    public void setNotaobtenida(Float notaobtenida) {
+    public void setNotaobtenida(BigDecimal notaobtenida) {
         this.notaobtenida = notaobtenida;
     }
     public Boolean getBorrado() {
